refactor(client): rename folder eComMerchant/ to client/
